Allow players to teleport onto the Sky Wave
This is still not feature complete. When a player is teleported onto the
Sky Wave, they are still viible to everyone else. They look like they're
standing on nothing, but to them they're on the Sky Wave.

This will be fixed in the future once I can come up with a way to
differentiate between Echo Ridge and the Wave World.