Refactor function that was doing too many things
Having the Player and Character being configured at the same time only
served to add unnessecary complexity.

Now you call configurePlayer in PlayerAdded and configureCharacter in
CharacterAdded. Now you can choose to ommit configuring the Player or
Character.

This also makes the example unecessary, as it was only there to clear up
where the function should be called.