Remove trans-out functionality
This was an unused feature and not fully integrated. A better solution
for leaving the Wave Road needs to be thought up.

Having an exit pad inside the Wave Road model down near the Wave Station
would make the Wave Road's region extend down to Echo Ridge. This is a
substantial issue, and until a better solution is in place transing out
of the Wave Road is being removed.