Handle showing the Wave Road server-side for a better user experience
This is both for the user and so I can actually complete this feature.

I initially intended on having it so that when a player enters a Wave Road,
it's only shown client-side. Then all players on the Wave Road are made
visible to you. People in Echo Ridge can not see anyone on the Wave Road,
or the Wave Road itself.

This turned out to be a very complicated solution. Instead when a player
enters the Wave Road, it is shown to everybody. This gives a visual cue
that there's more to explore, and it also lets me finally implement the
damn thing.