Remove the use of Nevermore
There's still a few things leftover in the main server script, but those
will be removed later after investigating if they were fixes for Nevermore
or ROBLOX in general.

All of the paths and references to nevermore and its functions has been
completely removed. Ran some tests and everything is working normally.
We're officially separated!