Fix the camera slightly bobbing along to the character
The character's PrimaryPart (its head) is animated, along with the rest of
the body, making the camera bob up and down slightly when the "breathing"
animation is playing.

It's hardly noticable, but definitly needs to be fixed. We're locking the
camera onto the HumanoidRootPart to fix this problem, as it remains static.