Replace current interaction code with triggers
Triggers are Parts that act similar to Region3s (they're just parts,
though). When a user is inside of one they're allowed to "interact" with
an object.

This really just fires an event to the server so that the server can take
care of what happens next, though. The user isn't actually interacting with
anything, they're just pressing a key to alert the server.

Known bugs:

- TouchEnded doesn't register that touching the trigger stopped when being
  teleporting onto the Sky Wave. This leads to the user still being able to
  use the interact keybind.