Use plain Region3s from a wrapper class instead of the Region module
This new Region module shares the same name as the last one, but it's
considerably different. There's much less math as we're basing all of our
regions off of the built in Region3s.

We don't need to consider rotation with regions at all, so this is a very
simple, significantly less mathy solution to our region based needs.