Resolve #2
Eventually we will need to reimplment this functionality, as changing
the Skybox and maybe ambient lighting will become a common occourance
soon enough.

For the time being, none of this code is necessary, and the way we
detect regions needs a rework anyway, so we'll revisit this at a
later date.