We don't need to keep the lsit of Trigger Parts updated
There is never a case where a Trigger Part will be added during the
game, so these events will never be used.

There may come a point in the future where Trigger Parts are added
during the course of the game, but for the time being this is unused
functionality and can be removed.