Turn BaseTrigger into an actual class
BaseTrigger was mostly just an abstract class with no real usage, but it's
now being turned into something we can use.

It's going to be simply for detecting Touched interactions, but we can
change that later if we need to. This class is mostly just to conceptualize
the idea of a "trigger" while still being useful on its own.

This also updates CharacterTrigger to work with the changes.