Overhaul Component detection
This is a really big change, but basically Components are now discovered by
a ComponentType StringValue, and Components that share the same type are
grouped together in a table.

We can now separate different types of Components (like Touch-based and
action-based Warps), life should be good now.

(God I hope there aren't gonna be more problems with Components)