Handle all teleportation client-side (resolves #5)
It's been a long time coming but we finally did it. This was such a pain to
implement, we needed a Warping system and a Component system which both took
about 2 weeks in total to get up and running.

But we finally did it, all teleportation is handled very cleanly now, and the
server has no part in it. That means everything is very snappy as we don't need
to wait for the server to allow the teleport.