Overhaul the Trigger classes
This is a really hefty update which comes with a few different changes:

Trigger now has a Whitelist so we can only allow select Parts to fire its
Touched event.

RegionTrigger is a new class which abstracts CharacterTrigger's Region
functionality.

CharacterTrigger is now a subclass of RegionTrigger, and all it does is add
the Character's HumanoidRootPart to the Whitelist.