Reimplement Cell and World handling (resolves #6)
When Warping, you are now moved between Cells if the Warp is a descendant
of a Cell Component.

Currently this doesn't do much, it's mostly just a foundation to build off
of later. We now can keep track of where each player is in the game, next
step is to be able to change the lighting, skybox, and camera.