Implement custom services
This has been a long time coming. After almost a month in the making we now have ModuleScript-based services, akin to ROBLOX's built-in services.

They're designed to be simple to use (require the module from client or server and you'r good to go) and handle all the icky server-client routing for you.

They're finally stable and we can start rolling them out. There's still more work to be done on them to clean up how we define services, but for now they're ready for their debut!