Remove LocalWorld and LocalCell classes
These were used as a convinience so we didn't have to pass in the player for each operation.

This ends up added unnecessary complication, along with API duplication as we have to keep LocalWorld and World, and LocalCell and Cell in sync.

Having API duplication, along with more files to deal with out weights the benefit of not having to pass in the client. It doesn't even take many morelines to do so.